add data to view, sign out
