+++ app/src/main/res/layout/fragment_profile.xml
@@ -228,6 +228,7 @@
             android:background="@color/gray_light" />
 
         <LinearLayout
+            android:id="@+id/ll_logout"
             android:layout_width="match_parent"
             android:layout_height="@dimen/dp_48"
             android:gravity="center_vertical"
